+++ library/core/src/test/java/com/google/android/exoplayer2/extractor/mp4/FragmentedMp4ExtractorTest.java
@@ -57,6 +57,12 @@ public final class FragmentedMp4ExtractorTest {
         getExtractorFactory(Collections.emptyList()), "mp4/sample_ac4_fragmented.mp4");
   }
 
+  @Test
+  public void testSampleWithProtectedAc4Track() throws Exception {
+    ExtractorAsserts.assertBehavior(
+        getExtractorFactory(Collections.emptyList()), "mp4/sample_ac4_protected.mp4");
+  }
+
   private static ExtractorFactory getExtractorFactory(final List<Format> closedCaptionFormats) {
     return () -> new FragmentedMp4Extractor(0, null, null, null, closedCaptionFormats);
   }