+++ RELEASENOTES.md
@@ -37,6 +37,7 @@
   on earlier releases, but only when embedded in a non-FLAC container such as
   Matroska or MP4.
 * Javadocs: Add favicon for easier identification in browser tabs
+* FMP4: Add support for encrypted AC-4 tracks.
 
 ### 2.11.1 (2019-12-20) ###